refactor(toolbar): extract FocusCycleManager from toolbar-ui.js into its own module (#7856)
* refactor(toolbar): extract FocusCycleManager from toolbar-ui.js into its own module


* refactor(toolbar): register activity/focus-cycle-manager in RequireJS loader deps


* fix(toolbar): make FocusCycleManager init idempotent and reattach live region on re-init


* test(toolbar): add unit tests for extracted FocusCycleManager module


* test(toolbar): assert listener add/remove symmetry and announcer singleton


---------
